requirements.txt
# =========================
# Core ML / Transformers
# =========================
--index-url https://download.pytorch.org/whl/cu130
--extra-index-url https://pypi.org/simple
torch
torchvision
torchaudio
transformers
sentencepiece

# =========================
# Image Enhancement & Vision
# =========================
-e ./temp_basicsr
realesrgan
gfpgan
opencv-python
scikit-image

# =========================
# Translation
# =========================
# (Using IndicTrans2, NLLB, MarianMT via transformers)

# =========================
# Terminal UI
# =========================
inquirer
tqdm
discord.py

# =========================
# Utilities
# =========================
requests
numpy
Pillow
python-dotenv
psutil

# =========================
# Python 3.14 Compatibility
# =========================
# spacy removed - not used in project (caused build errors on Windows)
